#1521 - Refactoring
